requirements.txt
requests
GitPython
pandas
pyyaml
